+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - New #362: Add `GridView::captionAttributes()` method and `$attributes` parameter to `GridView::caption()` to set
   HTML attributes for the `caption` tag (@vjik)
 - Enh #358: Make dependency container in `GridView` constructor optional (@vjik)
+- Enh #363: Add `scope="col"` attribute to `GridView` header cells by default (@vjik)
 
 ## 1.2.0 September 02, 2026